refactor: 공유모달 파일 위치 변경
