+++ src/components/ReportDetail/SharedModal.tsx
